[rename/#42] market -> store로 변경
