+++ feature/store/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 android {
-    setNameSpace("feature.market")
+    setNameSpace("feature.store")
 }
 
 dependencies {
