+++ feature/store/src/main/AndroidManifest.xml
