+++ feature/store/src/main/java/com/napzak/market/store/component/GenreChip.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.market.component
+package com.napzak.market.store.component
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
@@ -12,7 +12,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.napzak.market.designsystem.theme.NapzakMarketTheme
-import com.napzak.market.feature.market.R.string.market_ellipsis
+import com.napzak.market.feature.store.R.string.market_ellipsis
 
 private const val MAX_LENGTH = 5
