+++ feature/store/src/main/res/drawable/ic_right_chevron.xml
