+++ feature/store/src/main/res/values/strings.xml
